requirements.txt
email_validator==2.2.0
fastapi==0.111.1
openai==1.37.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
uvicorn==0.30.3
ultralytics==8.3.10
opencv-python==4.10.0.84
pillow==10.4.0